google_sheets/requirements.txt
celery
redis
gspread
google-auth
pydantic
pydantic-settings
apscheduler